requirements.txt
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.5
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pycodestyle==2.6.0
pycparser==2.20
pyheif==0.5.1
toml==0.10.2
whatimage==0.0.3
